+++ dns/transport/local/resolv_windows.go
@@ -5,15 +5,23 @@ import (
 	"net"
 	"net/netip"
 	"os"
+	"strings"
 	"strconv"
 	"syscall"
 	"time"
 	"unsafe"
 
 	"github.com/sagernet/sing-box/adapter"
+	"github.com/sagernet/sing/common"
 	"github.com/sagernet/sing/service"
 
 	"golang.org/x/sys/windows"
+)
+
+const (
+	localAddr    = "127.1.1.1"
+	dhcpMarkAddr = "127.1.2.3"
+	setMarkAddr  = "127.3.2.1"
 )
 
 func dnsReadConfig(ctx context.Context, _ string) *dnsConfig {
@@ -77,16 +85,30 @@ func dnsReadConfig(ctx context.Context, _ string) *dnsConfig {
 			}{ifName: windows.UTF16PtrToString(address.FriendlyName), Addr: dnsServerAddr})
 		}
 	}
-	var myInterface string
+	var myInterfaces []string
 	if networkManager := service.FromContext[adapter.NetworkManager](ctx); networkManager != nil {
-		myInterface = networkManager.InterfaceMonitor().MyInterface()
+		myInterfaces = networkManager.InterfaceMonitor().MyInterfaces()
 	}
+	customDns := false
 	for _, address := range dnsAddresses {
-		if address.ifName == myInterface {
+		if common.Contains(myInterfaces, address.ifName) {
+			continue
+		}
+		if address.String() == dhcpMarkAddr || address.String() == setMarkAddr {
+			customDns = true
 			continue
 		}
 		conf.servers = append(conf.servers, net.JoinHostPort(address.String(), "53"))
 	}
+	servers := make([]string, 0)
+	for _, server := range conf.servers {
+		addr := strings.Split(server, ":")[0]
+		if addr == localAddr && customDns {
+			continue
+		}
+		servers = append(servers, server)
+	}
+	conf.servers = servers
 	return conf
 }