+++ experimental/clashapi/server.go
@@ -63,7 +63,15 @@ type Server struct {
 }
 
 func NewServer(ctx context.Context, logFactory log.ObservableFactory, options option.ClashAPIOptions) (adapter.ClashServer, error) {
-	trafficManager := trafficontrol.NewManager()
+	outStr := make([]string, 0)
+	for _, out := range service.FromContext[adapter.OutboundManager](ctx).Outbounds() {
+		outStr = append(outStr, out.Tag())
+	}
+	endStr := make([]string, 0)
+	for _, end := range service.FromContext[adapter.EndpointManager](ctx).Endpoints() {
+		endStr = append(endStr, end.Tag())
+	}
+	trafficManager := trafficontrol.NewManager(outStr, endStr)
 	chiRouter := chi.NewRouter()
 	s := &Server{
 		ctx:       ctx,
@@ -164,7 +172,7 @@ func (s *Server) Start(stage adapter.StartStage) error {
 				listener net.Listener
 				err      error
 			)
-			for i := 0; i < 3; i++ {
+			for range 3 {
 				listener, err = net.Listen("tcp", s.httpServer.Addr)
 				if runtime.GOOS == "android" && errors.Is(err, syscall.EADDRINUSE) {
 					time.Sleep(100 * time.Millisecond)